move subprotocol interface to protocol
